Close outline by spreading gap

Distribute the distance between first and last points across all points to close the outline.

Usage

coo_close_spread(x, ..., .cols = NULL, .ldk_col = NULL)

Arguments

x

A matrix (nx2), list of matrices, or tibble with coo columns.

...

Additional arguments (reserved for future use).

.cols

Column name(s) to process when x is a tibble. If NULL, automatically detects columns containing coo columns.

.ldk_col

Character. Name of landmark column. If NULL, uses colname_ldk.

Value

If x is a single matrix: returns the closed matrix

If x is a list: returns a list of closed matrices

If x is a tibble: returns the tibble with closed coo column(s)

Details

For open outlines, distributes the gap between the first and last points evenly across all points, so the outline becomes closed.

Landmarks are synced to maintain their positions.

See also

Examples

coo_close_spread(shapes$cat)
#> $coo
#>            [,1]       [,2]
#>   [1,] 200.0000  62.000000
#>   [2,] 195.9832  55.974790
#>   [3,] 199.9664  55.949580
#>   [4,] 205.9496  53.924370
#>   [5,] 208.9328  47.899160
#>   [6,] 204.9160  42.873950
#>   [7,] 199.8992  42.848739
#>   [8,] 193.8824  41.823529
#>   [9,] 187.8655  41.798319
#>  [10,] 181.8487  41.773109
#>  [11,] 175.8319  42.747899
#>  [12,] 169.8151  42.722689
#>  [13,] 163.7983  42.697479
#>  [14,] 157.7815  42.672269
#>  [15,] 151.7647  42.647059
#>  [16,] 145.7479  43.621849
#>  [17,] 142.7311  40.596639
#>  [18,] 141.7143  35.571429
#>  [19,] 144.6975  29.546218
#>  [20,] 149.6807  24.521008
#>  [21,] 155.6639  22.495798
#>  [22,] 161.6471  20.470588
#>  [23,] 167.6303  19.445378
#>  [24,] 173.6134  18.420168
#>  [25,] 179.5966  18.394958
#>  [26,] 185.5798  19.369748
#>  [27,] 191.5630  20.344538
#>  [28,] 197.5462  21.319328
#>  [29,] 202.5294  21.294118
#>  [30,] 207.5126  18.268908
#>  [31,] 201.4958  14.243697
#>  [32,] 195.4790  12.218487
#>  [33,] 189.4622  10.193277
#>  [34,] 183.4454   9.168067
#>  [35,] 177.4286   8.142857
#>  [36,] 171.4118   8.117647
#>  [37,] 165.3950   8.092437
#>  [38,] 159.3782   9.067227
#>  [39,] 153.3613  10.042017
#>  [40,] 148.3445  12.016807
#>  [41,] 142.3277  14.991597
#>  [42,] 136.3109  19.966387
#>  [43,] 132.2941  25.941176
#>  [44,] 129.2773  31.915966
#>  [45,] 129.2605  37.890756
#>  [46,] 129.2437  43.865546
#>  [47,] 131.2269  49.840336
#>  [48,] 134.2101  55.815126
#>  [49,] 135.1933  61.789916
#>  [50,] 135.1765  67.764706
#>  [51,] 134.1597  72.739496
#>  [52,] 133.1429  78.714286
#>  [53,] 133.1261  84.689076
#>  [54,] 132.1092  90.663866
#>  [55,] 132.0924  96.638655
#>  [56,] 132.0756 102.613445
#>  [57,] 133.0588 108.588235
#>  [58,] 133.0420 114.563025
#>  [59,] 135.0252 120.537815
#>  [60,] 138.0084 126.512605
#>  [61,] 140.9916 132.487395
#>  [62,] 142.9748 137.462185
#>  [63,] 146.9580 143.436975
#>  [64,] 151.9412 149.411765
#>  [65,] 157.9244 155.386555
#>  [66,] 163.9076 158.361345
#>  [67,] 169.8908 160.336134
#>  [68,] 175.8739 163.310924
#>  [69,] 181.8571 168.285714
#>  [70,] 186.8403 174.260504
#>  [71,] 189.8235 180.235294
#>  [72,] 189.8067 185.210084
#>  [73,] 190.7899 191.184874
#>  [74,] 192.7731 197.159664
#>  [75,] 195.7563 203.134454
#>  [76,] 201.7395 208.109244
#>  [77,] 207.7227 212.084034
#>  [78,] 210.7059 218.058824
#>  [79,] 214.6891 224.033613
#>  [80,] 219.6723 229.008403
#>  [81,] 223.6555 223.983193
#>  [82,] 224.6387 217.957983
#>  [83,] 224.6218 212.932773
#>  [84,] 226.6050 208.907563
#>  [85,] 232.5882 205.882353
#>  [86,] 237.5714 201.857143
#>  [87,] 242.5546 196.831933
#>  [88,] 247.5378 192.806723
#>  [89,] 246.5210 186.781513
#>  [90,] 242.5042 180.756303
#>  [91,] 236.4874 175.731092
#>  [92,] 233.4706 170.705882
#>  [93,] 233.4538 164.680672
#>  [94,] 234.4370 159.655462
#>  [95,] 235.4202 153.630252
#>  [96,] 235.4034 147.605042
#>  [97,] 235.3866 141.579832
#>  [98,] 235.3697 135.554622
#>  [99,] 233.3529 129.529412
#> [100,] 229.3361 123.504202
#> [101,] 225.3193 117.478992
#> [102,] 222.3025 111.453782
#> [103,] 220.2857 105.428571
#> [104,] 219.2689  99.403361
#> [105,] 219.2521  94.378151
#> [106,] 219.2353  88.352941
#> [107,] 219.2185  82.327731
#> [108,] 219.2017  76.302521
#> [109,] 219.1849  70.277311
#> [110,] 220.1681  64.252101
#> [111,] 222.1513  59.226891
#> [112,] 228.1345  57.201681
#> [113,] 230.1176  52.176471
#> [114,] 228.1008  46.151261
#> [115,] 222.0840  42.126050
#> [116,] 217.0672  42.100840
#> [117,] 211.0504  44.075630
#> [118,] 206.0336  50.050420
#> [119,] 202.0168  56.025210
#> [120,] 200.0000  62.000000
#> 
#> $ldk
#> NULL
#> 
#> attr(,"class")
#> [1] "coo"  "list"