Extract or subset coordinates

Subset or trim shapes by row index or position.

Usage

coo_extract(x, ..., .cols = NULL, .ldk_col = NULL)

coo_head(x, ..., .cols = NULL, .ldk_col = NULL)

coo_tail(x, ..., .cols = NULL, .ldk_col = NULL)

Arguments

x

A matrix (nx2), list of matrices, or tibble with coo columns.

...

Additional arguments (reserved for future use).

.cols

Column name(s) to process when x is a tibble. If NULL, automatically detects columns containing coo objects.

.ldk_col

Character. Name of landmark column. If NULL, uses colname_ldk.

id

Integer vector of row indices to extract (for coo_extract()).

n

Integer. Number of points to keep or remove.

Value

If x is a single matrix: returns a matrix with selected/trimmed rows

If x is a list: returns a list of matrices

If x is a tibble: returns the tibble with specified coo columns subsetted

Details

coo_extract(): extract specific rows by index

coo_head(): keep first n points

coo_tail(): keep all but the last n points

If n equals nrow(x), the shape is returned unchanged.

These functions are landmark-aware: invalid landmark indices are automatically filtered out after subsetting.

See also

Examples

coo_extract(shapes$cat, c(1, 10, 50))
#>      [,1] [,2]
#> [1,]  200   62
#> [2,]  182   42
#> [3,]  136   69
coo_extract(shapes)
#> $cat
#>        [,1] [,2]
#>   [1,]  200   62
#>   [2,]  196   56
#>   [3,]  200   56
#>   [4,]  206   54
#>   [5,]  209   48
#>   [6,]  205   43
#>   [7,]  200   43
#>   [8,]  194   42
#>   [9,]  188   42
#>  [10,]  182   42
#>  [11,]  176   43
#>  [12,]  170   43
#>  [13,]  164   43
#>  [14,]  158   43
#>  [15,]  152   43
#>  [16,]  146   44
#>  [17,]  143   41
#>  [18,]  142   36
#>  [19,]  145   30
#>  [20,]  150   25
#>  [21,]  156   23
#>  [22,]  162   21
#>  [23,]  168   20
#>  [24,]  174   19
#>  [25,]  180   19
#>  [26,]  186   20
#>  [27,]  192   21
#>  [28,]  198   22
#>  [29,]  203   22
#>  [30,]  208   19
#>  [31,]  202   15
#>  [32,]  196   13
#>  [33,]  190   11
#>  [34,]  184   10
#>  [35,]  178    9
#>  [36,]  172    9
#>  [37,]  166    9
#>  [38,]  160   10
#>  [39,]  154   11
#>  [40,]  149   13
#>  [41,]  143   16
#>  [42,]  137   21
#>  [43,]  133   27
#>  [44,]  130   33
#>  [45,]  130   39
#>  [46,]  130   45
#>  [47,]  132   51
#>  [48,]  135   57
#>  [49,]  136   63
#>  [50,]  136   69
#>  [51,]  135   74
#>  [52,]  134   80
#>  [53,]  134   86
#>  [54,]  133   92
#>  [55,]  133   98
#>  [56,]  133  104
#>  [57,]  134  110
#>  [58,]  134  116
#>  [59,]  136  122
#>  [60,]  139  128
#>  [61,]  142  134
#>  [62,]  144  139
#>  [63,]  148  145
#>  [64,]  153  151
#>  [65,]  159  157
#>  [66,]  165  160
#>  [67,]  171  162
#>  [68,]  177  165
#>  [69,]  183  170
#>  [70,]  188  176
#>  [71,]  191  182
#>  [72,]  191  187
#>  [73,]  192  193
#>  [74,]  194  199
#>  [75,]  197  205
#>  [76,]  203  210
#>  [77,]  209  214
#>  [78,]  212  220
#>  [79,]  216  226
#>  [80,]  221  231
#>  [81,]  225  226
#>  [82,]  226  220
#>  [83,]  226  215
#>  [84,]  228  211
#>  [85,]  234  208
#>  [86,]  239  204
#>  [87,]  244  199
#>  [88,]  249  195
#>  [89,]  248  189
#>  [90,]  244  183
#>  [91,]  238  178
#>  [92,]  235  173
#>  [93,]  235  167
#>  [94,]  236  162
#>  [95,]  237  156
#>  [96,]  237  150
#>  [97,]  237  144
#>  [98,]  237  138
#>  [99,]  235  132
#> [100,]  231  126
#> [101,]  227  120
#> [102,]  224  114
#> [103,]  222  108
#> [104,]  221  102
#> [105,]  221   97
#> [106,]  221   91
#> [107,]  221   85
#> [108,]  221   79
#> [109,]  221   73
#> [110,]  222   67
#> [111,]  224   62
#> [112,]  230   60
#> [113,]  232   55
#> [114,]  230   49
#> [115,]  224   45
#> [116,]  219   45
#> [117,]  213   47
#> [118,]  208   53
#> [119,]  204   59
#> [120,]  202   65
#> 
#> $dog
#>        [,1] [,2]
#>   [1,]  200  106
#>   [2,]  194  105
#>   [3,]  187  106
#>   [4,]  181  106
#>   [5,]  174  107
#>   [6,]  168  108
#>   [7,]  162  107
#>   [8,]  156  100
#>   [9,]  152   94
#>  [10,]  145   87
#>  [11,]  139   83
#>  [12,]  133   78
#>  [13,]  127   71
#>  [14,]  124   65
#>  [15,]  122   59
#>  [16,]  126   52
#>  [17,]  128   47
#>  [18,]  121   45
#>  [19,]  115   45
#>  [20,]  110   49
#>  [21,]  109   56
#>  [22,]  109   62
#>  [23,]  109   69
#>  [24,]  109   75
#>  [25,]  112   81
#>  [26,]  119   88
#>  [27,]  124   94
#>  [28,]  126  101
#>  [29,]  123   96
#>  [30,]  117   90
#>  [31,]  110   85
#>  [32,]  104   82
#>  [33,]  107   89
#>  [34,]  112   95
#>  [35,]  113  101
#>  [36,]  115  108
#>  [37,]  115  114
#>  [38,]  116  120
#>  [39,]  115  127
#>  [40,]  116  133
#>  [41,]  117  140
#>  [42,]  120  146
#>  [43,]  124  152
#>  [44,]  131  159
#>  [45,]  137  162
#>  [46,]  144  164
#>  [47,]  150  165
#>  [48,]  156  165
#>  [49,]  163  166
#>  [50,]  169  166
#>  [51,]  176  166
#>  [52,]  182  166
#>  [53,]  188  166
#>  [54,]  195  165
#>  [55,]  201  165
#>  [56,]  208  166
#>  [57,]  214  166
#>  [58,]  220  168
#>  [59,]  227  170
#>  [60,]  233  174
#>  [61,]  238  179
#>  [62,]  243  186
#>  [63,]  247  192
#>  [64,]  251  199
#>  [65,]  253  205
#>  [66,]  256  211
#>  [67,]  258  218
#>  [68,]  264  224
#>  [69,]  271  227
#>  [70,]  277  227
#>  [71,]  283  227
#>  [72,]  290  222
#>  [73,]  295  216
#>  [74,]  302  214
#>  [75,]  308  214
#>  [76,]  314  214
#>  [77,]  316  207
#>  [78,]  316  201
#>  [79,]  315  194
#>  [80,]  310  190
#>  [81,]  304  188
#>  [82,]  297  188
#>  [83,]  291  186
#>  [84,]  289  179
#>  [85,]  286  173
#>  [86,]  284  167
#>  [87,]  281  160
#>  [88,]  279  154
#>  [89,]  277  148
#>  [90,]  277  141
#>  [91,]  275  135
#>  [92,]  274  128
#>  [93,]  273  122
#>  [94,]  270  116
#>  [95,]  267  109
#>  [96,]  266  103
#>  [97,]  266   96
#>  [98,]  266   90
#>  [99,]  266   84
#> [100,]  267   77
#> [101,]  267   71
#> [102,]  268   64
#> [103,]  274   60
#> [104,]  277   55
#> [105,]  270   51
#> [106,]  264   50
#> [107,]  259   55
#> [108,]  258   61
#> [109,]  257   67
#> [110,]  255   74
#> [111,]  252   80
#> [112,]  250   86
#> [113,]  247   93
#> [114,]  244   99
#> [115,]  238  104
#> [116,]  232  105
#> [117,]  226  105
#> [118,]  219  105
#> [119,]  213  106
#> [120,]  206  106
#> 
#> $heart
#>        [,1] [,2]
#>   [1,]  200   37
#>   [2,]  198   39
#>   [3,]  197   42
#>   [4,]  193   47
#>   [5,]  190   51
#>   [6,]  185   56
#>   [7,]  180   61
#>   [8,]  176   65
#>   [9,]  171   69
#>  [10,]  167   72
#>  [11,]  162   76
#>  [12,]  157   79
#>  [13,]  153   82
#>  [14,]  148   85
#>  [15,]  143   88
#>  [16,]  139   90
#>  [17,]  134   93
#>  [18,]  130   96
#>  [19,]  125  100
#>  [20,]  120  104
#>  [21,]  116  108
#>  [22,]  111  113
#>  [23,]  108  117
#>  [24,]  105  122
#>  [25,]  103  127
#>  [26,]  101  131
#>  [27,]  100  136
#>  [28,]   99  140
#>  [29,]   99  145
#>  [30,]   99  150
#>  [31,]   99  154
#>  [32,]   99  159
#>  [33,]   99  163
#>  [34,]  100  168
#>  [35,]  102  173
#>  [36,]  103  177
#>  [37,]  105  182
#>  [38,]  108  187
#>  [39,]  111  191
#>  [40,]  114  196
#>  [41,]  118  200
#>  [42,]  123  204
#>  [43,]  128  207
#>  [44,]  132  209
#>  [45,]  137  211
#>  [46,]  141  212
#>  [47,]  146  213
#>  [48,]  151  213
#>  [49,]  155  213
#>  [50,]  160  212
#>  [51,]  164  211
#>  [52,]  169  209
#>  [53,]  174  207
#>  [54,]  178  204
#>  [55,]  183  200
#>  [56,]  187  197
#>  [57,]  191  192
#>  [58,]  195  187
#>  [59,]  198  183
#>  [60,]  200  178
#>  [61,]  202  175
#>  [62,]  204  179
#>  [63,]  207  184
#>  [64,]  210  188
#>  [65,]  214  193
#>  [66,]  219  198
#>  [67,]  223  202
#>  [68,]  228  205
#>  [69,]  232  207
#>  [70,]  237  209
#>  [71,]  242  211
#>  [72,]  246  211
#>  [73,]  251  212
#>  [74,]  255  212
#>  [75,]  260  212
#>  [76,]  265  211
#>  [77,]  269  210
#>  [78,]  274  207
#>  [79,]  278  205
#>  [80,]  283  202
#>  [81,]  288  197
#>  [82,]  292  193
#>  [83,]  296  188
#>  [84,]  298  184
#>  [85,]  300  179
#>  [86,]  302  174
#>  [87,]  303  170
#>  [88,]  304  165
#>  [89,]  305  160
#>  [90,]  306  156
#>  [91,]  305  151
#>  [92,]  305  147
#>  [93,]  305  142
#>  [94,]  305  137
#>  [95,]  303  133
#>  [96,]  302  128
#>  [97,]  300  124
#>  [98,]  297  119
#>  [99,]  293  114
#> [100,]  289  110
#> [101,]  284  105
#> [102,]  280  102
#> [103,]  275   98
#> [104,]  270   94
#> [105,]  266   92
#> [106,]  261   89
#> [107,]  257   86
#> [108,]  252   83
#> [109,]  247   80
#> [110,]  243   77
#> [111,]  238   73
#> [112,]  233   69
#> [113,]  229   65
#> [114,]  224   61
#> [115,]  220   57
#> [116,]  215   52
#> [117,]  211   47
#> [118,]  208   43
#> [119,]  205   38
#> [120,]  203   34
#> 
#> $leaf2
#>        [,1] [,2]
#>   [1,]  200   62
#>   [2,]  200   55
#>   [3,]  199   49
#>   [4,]  197   42
#>   [5,]  196   36
#>   [6,]  194   29
#>   [7,]  190   22
#>   [8,]  184   20
#>   [9,]  179   24
#>  [10,]  183   30
#>  [11,]  187   37
#>  [12,]  189   44
#>  [13,]  190   50
#>  [14,]  191   57
#>  [15,]  192   63
#>  [16,]  192   70
#>  [17,]  186   68
#>  [18,]  179   66
#>  [19,]  172   65
#>  [20,]  166   65
#>  [21,]  159   66
#>  [22,]  153   66
#>  [23,]  146   68
#>  [24,]  139   70
#>  [25,]  133   72
#>  [26,]  126   75
#>  [27,]  127   80
#>  [28,]  133   87
#>  [29,]  140   94
#>  [30,]  142   99
#>  [31,]  135  103
#>  [32,]  129  107
#>  [33,]  122  112
#>  [34,]  115  119
#>  [35,]  116  124
#>  [36,]  123  127
#>  [37,]  127  129
#>  [38,]  122  136
#>  [39,]  117  142
#>  [40,]  113  149
#>  [41,]  109  156
#>  [42,]  109  162
#>  [43,]  116  162
#>  [44,]  122  161
#>  [45,]  129  160
#>  [46,]  133  161
#>  [47,]  134  167
#>  [48,]  139  171
#>  [49,]  145  168
#>  [50,]  152  163
#>  [51,]  159  158
#>  [52,]  165  152
#>  [53,]  166  153
#>  [54,]  164  159
#>  [55,]  162  166
#>  [56,]  160  173
#>  [57,]  159  179
#>  [58,]  159  186
#>  [59,]  161  192
#>  [60,]  168  191
#>  [61,]  174  189
#>  [62,]  181  187
#>  [63,]  181  194
#>  [64,]  182  200
#>  [65,]  184  207
#>  [66,]  186  213
#>  [67,]  190  220
#>  [68,]  196  226
#>  [69,]  202  221
#>  [70,]  206  214
#>  [71,]  208  208
#>  [72,]  210  201
#>  [73,]  210  194
#>  [74,]  211  188
#>  [75,]  217  189
#>  [76,]  223  191
#>  [77,]  230  192
#>  [78,]  233  186
#>  [79,]  233  180
#>  [80,]  232  173
#>  [81,]  231  167
#>  [82,]  229  160
#>  [83,]  226  153
#>  [84,]  226  151
#>  [85,]  233  158
#>  [86,]  239  163
#>  [87,]  246  168
#>  [88,]  252  171
#>  [89,]  258  168
#>  [90,]  259  161
#>  [91,]  262  159
#>  [92,]  269  161
#>  [93,]  275  162
#>  [94,]  282  162
#>  [95,]  282  156
#>  [96,]  280  150
#>  [97,]  275  143
#>  [98,]  271  137
#>  [99,]  266  130
#> [100,]  269  127
#> [101,]  275  125
#> [102,]  276  119
#> [103,]  271  113
#> [104,]  264  107
#> [105,]  257  103
#> [106,]  251   99
#> [107,]  252   94
#> [108,]  258   88
#> [109,]  264   81
#> [110,]  267   75
#> [111,]  260   72
#> [112,]  253   70
#> [113,]  247   68
#> [114,]  240   66
#> [115,]  234   66
#> [116,]  227   65
#> [117,]  220   65
#> [118,]  214   66
#> [119,]  207   68
#> [120,]  201   69
#> 
coo_head(shapes$cat, 10)
#>       [,1] [,2]
#>  [1,]  200   62
#>  [2,]  196   56
#>  [3,]  200   56
#>  [4,]  206   54
#>  [5,]  209   48
#>  [6,]  205   43
#>  [7,]  200   43
#>  [8,]  194   42
#>  [9,]  188   42
#> [10,]  182   42
coo_head(shapes)
#> $cat
#>       [,1] [,2]
#>  [1,]  200   62
#>  [2,]  196   56
#>  [3,]  200   56
#>  [4,]  206   54
#>  [5,]  209   48
#>  [6,]  205   43
#>  [7,]  200   43
#>  [8,]  194   42
#>  [9,]  188   42
#> [10,]  182   42
#> 
#> $dog
#>       [,1] [,2]
#>  [1,]  200  106
#>  [2,]  194  105
#>  [3,]  187  106
#>  [4,]  181  106
#>  [5,]  174  107
#>  [6,]  168  108
#>  [7,]  162  107
#>  [8,]  156  100
#>  [9,]  152   94
#> [10,]  145   87
#> 
#> $heart
#>       [,1] [,2]
#>  [1,]  200   37
#>  [2,]  198   39
#>  [3,]  197   42
#>  [4,]  193   47
#>  [5,]  190   51
#>  [6,]  185   56
#>  [7,]  180   61
#>  [8,]  176   65
#>  [9,]  171   69
#> [10,]  167   72
#> 
#> $leaf2
#>       [,1] [,2]
#>  [1,]  200   62
#>  [2,]  200   55
#>  [3,]  199   49
#>  [4,]  197   42
#>  [5,]  196   36
#>  [6,]  194   29
#>  [7,]  190   22
#>  [8,]  184   20
#>  [9,]  179   24
#> [10,]  183   30
#> 
coo_tail(shapes$cat, 5)
#>        [,1] [,2]
#>   [1,]  200   62
#>   [2,]  196   56
#>   [3,]  200   56
#>   [4,]  206   54
#>   [5,]  209   48
#>   [6,]  205   43
#>   [7,]  200   43
#>   [8,]  194   42
#>   [9,]  188   42
#>  [10,]  182   42
#>  [11,]  176   43
#>  [12,]  170   43
#>  [13,]  164   43
#>  [14,]  158   43
#>  [15,]  152   43
#>  [16,]  146   44
#>  [17,]  143   41
#>  [18,]  142   36
#>  [19,]  145   30
#>  [20,]  150   25
#>  [21,]  156   23
#>  [22,]  162   21
#>  [23,]  168   20
#>  [24,]  174   19
#>  [25,]  180   19
#>  [26,]  186   20
#>  [27,]  192   21
#>  [28,]  198   22
#>  [29,]  203   22
#>  [30,]  208   19
#>  [31,]  202   15
#>  [32,]  196   13
#>  [33,]  190   11
#>  [34,]  184   10
#>  [35,]  178    9
#>  [36,]  172    9
#>  [37,]  166    9
#>  [38,]  160   10
#>  [39,]  154   11
#>  [40,]  149   13
#>  [41,]  143   16
#>  [42,]  137   21
#>  [43,]  133   27
#>  [44,]  130   33
#>  [45,]  130   39
#>  [46,]  130   45
#>  [47,]  132   51
#>  [48,]  135   57
#>  [49,]  136   63
#>  [50,]  136   69
#>  [51,]  135   74
#>  [52,]  134   80
#>  [53,]  134   86
#>  [54,]  133   92
#>  [55,]  133   98
#>  [56,]  133  104
#>  [57,]  134  110
#>  [58,]  134  116
#>  [59,]  136  122
#>  [60,]  139  128
#>  [61,]  142  134
#>  [62,]  144  139
#>  [63,]  148  145
#>  [64,]  153  151
#>  [65,]  159  157
#>  [66,]  165  160
#>  [67,]  171  162
#>  [68,]  177  165
#>  [69,]  183  170
#>  [70,]  188  176
#>  [71,]  191  182
#>  [72,]  191  187
#>  [73,]  192  193
#>  [74,]  194  199
#>  [75,]  197  205
#>  [76,]  203  210
#>  [77,]  209  214
#>  [78,]  212  220
#>  [79,]  216  226
#>  [80,]  221  231
#>  [81,]  225  226
#>  [82,]  226  220
#>  [83,]  226  215
#>  [84,]  228  211
#>  [85,]  234  208
#>  [86,]  239  204
#>  [87,]  244  199
#>  [88,]  249  195
#>  [89,]  248  189
#>  [90,]  244  183
#>  [91,]  238  178
#>  [92,]  235  173
#>  [93,]  235  167
#>  [94,]  236  162
#>  [95,]  237  156
#>  [96,]  237  150
#>  [97,]  237  144
#>  [98,]  237  138
#>  [99,]  235  132
#> [100,]  231  126
#> [101,]  227  120
#> [102,]  224  114
#> [103,]  222  108
#> [104,]  221  102
#> [105,]  221   97
#> [106,]  221   91
#> [107,]  221   85
#> [108,]  221   79
#> [109,]  221   73
#> [110,]  222   67
#> [111,]  224   62
#> [112,]  230   60
#> [113,]  232   55
#> [114,]  230   49
#> [115,]  224   45
coo_head(bot, 20)
#> # A tibble: 40 × 3
#>    coo            type   dummy
#>    <named list>   <fct>  <fct>
#>  1 <dbl [20 × 2]> whisky a    
#>  2 <dbl [20 × 2]> whisky a    
#>  3 <dbl [20 × 2]> whisky a    
#>  4 <dbl [20 × 2]> whisky a    
#>  5 <dbl [20 × 2]> whisky a    
#>  6 <dbl [20 × 2]> whisky a    
#>  7 <dbl [20 × 2]> whisky a    
#>  8 <dbl [20 × 2]> whisky a    
#>  9 <dbl [20 × 2]> whisky a    
#> 10 <dbl [20 × 2]> whisky a    
#> # ℹ 30 more rows