Reverse point order

Reverse the order of points along an outline.

Usage

coo_reverse(x, ..., .cols = NULL, .ldk_col = NULL)

Arguments

x

A matrix (nx2), list of matrices, or tibble with coo columns.

...

Additional arguments (reserved for future use).

.cols

Column name(s) to process when x is a tibble. If NULL, automatically detects columns containing coo objects.

.ldk_col

Character. Name of landmark column. If NULL, uses colname_ldk.

Value

If x is a single matrix: returns the matrix with reversed point order

If x is a list: returns a list of matrices with reversed point order

If x is a tibble: returns the tibble with specified coo columns reversed

Details

This function is landmark-aware: landmark indices are automatically reversed to match the new point order using the formula n - ldk + 1.

Examples

coo_reverse(shapes$cat)
#>        [,1] [,2]
#>   [1,]  202   65
#>   [2,]  204   59
#>   [3,]  208   53
#>   [4,]  213   47
#>   [5,]  219   45
#>   [6,]  224   45
#>   [7,]  230   49
#>   [8,]  232   55
#>   [9,]  230   60
#>  [10,]  224   62
#>  [11,]  222   67
#>  [12,]  221   73
#>  [13,]  221   79
#>  [14,]  221   85
#>  [15,]  221   91
#>  [16,]  221   97
#>  [17,]  221  102
#>  [18,]  222  108
#>  [19,]  224  114
#>  [20,]  227  120
#>  [21,]  231  126
#>  [22,]  235  132
#>  [23,]  237  138
#>  [24,]  237  144
#>  [25,]  237  150
#>  [26,]  237  156
#>  [27,]  236  162
#>  [28,]  235  167
#>  [29,]  235  173
#>  [30,]  238  178
#>  [31,]  244  183
#>  [32,]  248  189
#>  [33,]  249  195
#>  [34,]  244  199
#>  [35,]  239  204
#>  [36,]  234  208
#>  [37,]  228  211
#>  [38,]  226  215
#>  [39,]  226  220
#>  [40,]  225  226
#>  [41,]  221  231
#>  [42,]  216  226
#>  [43,]  212  220
#>  [44,]  209  214
#>  [45,]  203  210
#>  [46,]  197  205
#>  [47,]  194  199
#>  [48,]  192  193
#>  [49,]  191  187
#>  [50,]  191  182
#>  [51,]  188  176
#>  [52,]  183  170
#>  [53,]  177  165
#>  [54,]  171  162
#>  [55,]  165  160
#>  [56,]  159  157
#>  [57,]  153  151
#>  [58,]  148  145
#>  [59,]  144  139
#>  [60,]  142  134
#>  [61,]  139  128
#>  [62,]  136  122
#>  [63,]  134  116
#>  [64,]  134  110
#>  [65,]  133  104
#>  [66,]  133   98
#>  [67,]  133   92
#>  [68,]  134   86
#>  [69,]  134   80
#>  [70,]  135   74
#>  [71,]  136   69
#>  [72,]  136   63
#>  [73,]  135   57
#>  [74,]  132   51
#>  [75,]  130   45
#>  [76,]  130   39
#>  [77,]  130   33
#>  [78,]  133   27
#>  [79,]  137   21
#>  [80,]  143   16
#>  [81,]  149   13
#>  [82,]  154   11
#>  [83,]  160   10
#>  [84,]  166    9
#>  [85,]  172    9
#>  [86,]  178    9
#>  [87,]  184   10
#>  [88,]  190   11
#>  [89,]  196   13
#>  [90,]  202   15
#>  [91,]  208   19
#>  [92,]  203   22
#>  [93,]  198   22
#>  [94,]  192   21
#>  [95,]  186   20
#>  [96,]  180   19
#>  [97,]  174   19
#>  [98,]  168   20
#>  [99,]  162   21
#> [100,]  156   23
#> [101,]  150   25
#> [102,]  145   30
#> [103,]  142   36
#> [104,]  143   41
#> [105,]  146   44
#> [106,]  152   43
#> [107,]  158   43
#> [108,]  164   43
#> [109,]  170   43
#> [110,]  176   43
#> [111,]  182   42
#> [112,]  188   42
#> [113,]  194   42
#> [114,]  200   43
#> [115,]  205   43
#> [116,]  209   48
#> [117,]  206   54
#> [118,]  200   56
#> [119,]  196   56
#> [120,]  200   62
coo_reverse(shapes)
#> $cat
#>        [,1] [,2]
#>   [1,]  202   65
#>   [2,]  204   59
#>   [3,]  208   53
#>   [4,]  213   47
#>   [5,]  219   45
#>   [6,]  224   45
#>   [7,]  230   49
#>   [8,]  232   55
#>   [9,]  230   60
#>  [10,]  224   62
#>  [11,]  222   67
#>  [12,]  221   73
#>  [13,]  221   79
#>  [14,]  221   85
#>  [15,]  221   91
#>  [16,]  221   97
#>  [17,]  221  102
#>  [18,]  222  108
#>  [19,]  224  114
#>  [20,]  227  120
#>  [21,]  231  126
#>  [22,]  235  132
#>  [23,]  237  138
#>  [24,]  237  144
#>  [25,]  237  150
#>  [26,]  237  156
#>  [27,]  236  162
#>  [28,]  235  167
#>  [29,]  235  173
#>  [30,]  238  178
#>  [31,]  244  183
#>  [32,]  248  189
#>  [33,]  249  195
#>  [34,]  244  199
#>  [35,]  239  204
#>  [36,]  234  208
#>  [37,]  228  211
#>  [38,]  226  215
#>  [39,]  226  220
#>  [40,]  225  226
#>  [41,]  221  231
#>  [42,]  216  226
#>  [43,]  212  220
#>  [44,]  209  214
#>  [45,]  203  210
#>  [46,]  197  205
#>  [47,]  194  199
#>  [48,]  192  193
#>  [49,]  191  187
#>  [50,]  191  182
#>  [51,]  188  176
#>  [52,]  183  170
#>  [53,]  177  165
#>  [54,]  171  162
#>  [55,]  165  160
#>  [56,]  159  157
#>  [57,]  153  151
#>  [58,]  148  145
#>  [59,]  144  139
#>  [60,]  142  134
#>  [61,]  139  128
#>  [62,]  136  122
#>  [63,]  134  116
#>  [64,]  134  110
#>  [65,]  133  104
#>  [66,]  133   98
#>  [67,]  133   92
#>  [68,]  134   86
#>  [69,]  134   80
#>  [70,]  135   74
#>  [71,]  136   69
#>  [72,]  136   63
#>  [73,]  135   57
#>  [74,]  132   51
#>  [75,]  130   45
#>  [76,]  130   39
#>  [77,]  130   33
#>  [78,]  133   27
#>  [79,]  137   21
#>  [80,]  143   16
#>  [81,]  149   13
#>  [82,]  154   11
#>  [83,]  160   10
#>  [84,]  166    9
#>  [85,]  172    9
#>  [86,]  178    9
#>  [87,]  184   10
#>  [88,]  190   11
#>  [89,]  196   13
#>  [90,]  202   15
#>  [91,]  208   19
#>  [92,]  203   22
#>  [93,]  198   22
#>  [94,]  192   21
#>  [95,]  186   20
#>  [96,]  180   19
#>  [97,]  174   19
#>  [98,]  168   20
#>  [99,]  162   21
#> [100,]  156   23
#> [101,]  150   25
#> [102,]  145   30
#> [103,]  142   36
#> [104,]  143   41
#> [105,]  146   44
#> [106,]  152   43
#> [107,]  158   43
#> [108,]  164   43
#> [109,]  170   43
#> [110,]  176   43
#> [111,]  182   42
#> [112,]  188   42
#> [113,]  194   42
#> [114,]  200   43
#> [115,]  205   43
#> [116,]  209   48
#> [117,]  206   54
#> [118,]  200   56
#> [119,]  196   56
#> [120,]  200   62
#> 
#> $dog
#>        [,1] [,2]
#>   [1,]  206  106
#>   [2,]  213  106
#>   [3,]  219  105
#>   [4,]  226  105
#>   [5,]  232  105
#>   [6,]  238  104
#>   [7,]  244   99
#>   [8,]  247   93
#>   [9,]  250   86
#>  [10,]  252   80
#>  [11,]  255   74
#>  [12,]  257   67
#>  [13,]  258   61
#>  [14,]  259   55
#>  [15,]  264   50
#>  [16,]  270   51
#>  [17,]  277   55
#>  [18,]  274   60
#>  [19,]  268   64
#>  [20,]  267   71
#>  [21,]  267   77
#>  [22,]  266   84
#>  [23,]  266   90
#>  [24,]  266   96
#>  [25,]  266  103
#>  [26,]  267  109
#>  [27,]  270  116
#>  [28,]  273  122
#>  [29,]  274  128
#>  [30,]  275  135
#>  [31,]  277  141
#>  [32,]  277  148
#>  [33,]  279  154
#>  [34,]  281  160
#>  [35,]  284  167
#>  [36,]  286  173
#>  [37,]  289  179
#>  [38,]  291  186
#>  [39,]  297  188
#>  [40,]  304  188
#>  [41,]  310  190
#>  [42,]  315  194
#>  [43,]  316  201
#>  [44,]  316  207
#>  [45,]  314  214
#>  [46,]  308  214
#>  [47,]  302  214
#>  [48,]  295  216
#>  [49,]  290  222
#>  [50,]  283  227
#>  [51,]  277  227
#>  [52,]  271  227
#>  [53,]  264  224
#>  [54,]  258  218
#>  [55,]  256  211
#>  [56,]  253  205
#>  [57,]  251  199
#>  [58,]  247  192
#>  [59,]  243  186
#>  [60,]  238  179
#>  [61,]  233  174
#>  [62,]  227  170
#>  [63,]  220  168
#>  [64,]  214  166
#>  [65,]  208  166
#>  [66,]  201  165
#>  [67,]  195  165
#>  [68,]  188  166
#>  [69,]  182  166
#>  [70,]  176  166
#>  [71,]  169  166
#>  [72,]  163  166
#>  [73,]  156  165
#>  [74,]  150  165
#>  [75,]  144  164
#>  [76,]  137  162
#>  [77,]  131  159
#>  [78,]  124  152
#>  [79,]  120  146
#>  [80,]  117  140
#>  [81,]  116  133
#>  [82,]  115  127
#>  [83,]  116  120
#>  [84,]  115  114
#>  [85,]  115  108
#>  [86,]  113  101
#>  [87,]  112   95
#>  [88,]  107   89
#>  [89,]  104   82
#>  [90,]  110   85
#>  [91,]  117   90
#>  [92,]  123   96
#>  [93,]  126  101
#>  [94,]  124   94
#>  [95,]  119   88
#>  [96,]  112   81
#>  [97,]  109   75
#>  [98,]  109   69
#>  [99,]  109   62
#> [100,]  109   56
#> [101,]  110   49
#> [102,]  115   45
#> [103,]  121   45
#> [104,]  128   47
#> [105,]  126   52
#> [106,]  122   59
#> [107,]  124   65
#> [108,]  127   71
#> [109,]  133   78
#> [110,]  139   83
#> [111,]  145   87
#> [112,]  152   94
#> [113,]  156  100
#> [114,]  162  107
#> [115,]  168  108
#> [116,]  174  107
#> [117,]  181  106
#> [118,]  187  106
#> [119,]  194  105
#> [120,]  200  106
#> 
#> $heart
#>        [,1] [,2]
#>   [1,]  203   34
#>   [2,]  205   38
#>   [3,]  208   43
#>   [4,]  211   47
#>   [5,]  215   52
#>   [6,]  220   57
#>   [7,]  224   61
#>   [8,]  229   65
#>   [9,]  233   69
#>  [10,]  238   73
#>  [11,]  243   77
#>  [12,]  247   80
#>  [13,]  252   83
#>  [14,]  257   86
#>  [15,]  261   89
#>  [16,]  266   92
#>  [17,]  270   94
#>  [18,]  275   98
#>  [19,]  280  102
#>  [20,]  284  105
#>  [21,]  289  110
#>  [22,]  293  114
#>  [23,]  297  119
#>  [24,]  300  124
#>  [25,]  302  128
#>  [26,]  303  133
#>  [27,]  305  137
#>  [28,]  305  142
#>  [29,]  305  147
#>  [30,]  305  151
#>  [31,]  306  156
#>  [32,]  305  160
#>  [33,]  304  165
#>  [34,]  303  170
#>  [35,]  302  174
#>  [36,]  300  179
#>  [37,]  298  184
#>  [38,]  296  188
#>  [39,]  292  193
#>  [40,]  288  197
#>  [41,]  283  202
#>  [42,]  278  205
#>  [43,]  274  207
#>  [44,]  269  210
#>  [45,]  265  211
#>  [46,]  260  212
#>  [47,]  255  212
#>  [48,]  251  212
#>  [49,]  246  211
#>  [50,]  242  211
#>  [51,]  237  209
#>  [52,]  232  207
#>  [53,]  228  205
#>  [54,]  223  202
#>  [55,]  219  198
#>  [56,]  214  193
#>  [57,]  210  188
#>  [58,]  207  184
#>  [59,]  204  179
#>  [60,]  202  175
#>  [61,]  200  178
#>  [62,]  198  183
#>  [63,]  195  187
#>  [64,]  191  192
#>  [65,]  187  197
#>  [66,]  183  200
#>  [67,]  178  204
#>  [68,]  174  207
#>  [69,]  169  209
#>  [70,]  164  211
#>  [71,]  160  212
#>  [72,]  155  213
#>  [73,]  151  213
#>  [74,]  146  213
#>  [75,]  141  212
#>  [76,]  137  211
#>  [77,]  132  209
#>  [78,]  128  207
#>  [79,]  123  204
#>  [80,]  118  200
#>  [81,]  114  196
#>  [82,]  111  191
#>  [83,]  108  187
#>  [84,]  105  182
#>  [85,]  103  177
#>  [86,]  102  173
#>  [87,]  100  168
#>  [88,]   99  163
#>  [89,]   99  159
#>  [90,]   99  154
#>  [91,]   99  150
#>  [92,]   99  145
#>  [93,]   99  140
#>  [94,]  100  136
#>  [95,]  101  131
#>  [96,]  103  127
#>  [97,]  105  122
#>  [98,]  108  117
#>  [99,]  111  113
#> [100,]  116  108
#> [101,]  120  104
#> [102,]  125  100
#> [103,]  130   96
#> [104,]  134   93
#> [105,]  139   90
#> [106,]  143   88
#> [107,]  148   85
#> [108,]  153   82
#> [109,]  157   79
#> [110,]  162   76
#> [111,]  167   72
#> [112,]  171   69
#> [113,]  176   65
#> [114,]  180   61
#> [115,]  185   56
#> [116,]  190   51
#> [117,]  193   47
#> [118,]  197   42
#> [119,]  198   39
#> [120,]  200   37
#> 
#> $leaf2
#>        [,1] [,2]
#>   [1,]  201   69
#>   [2,]  207   68
#>   [3,]  214   66
#>   [4,]  220   65
#>   [5,]  227   65
#>   [6,]  234   66
#>   [7,]  240   66
#>   [8,]  247   68
#>   [9,]  253   70
#>  [10,]  260   72
#>  [11,]  267   75
#>  [12,]  264   81
#>  [13,]  258   88
#>  [14,]  252   94
#>  [15,]  251   99
#>  [16,]  257  103
#>  [17,]  264  107
#>  [18,]  271  113
#>  [19,]  276  119
#>  [20,]  275  125
#>  [21,]  269  127
#>  [22,]  266  130
#>  [23,]  271  137
#>  [24,]  275  143
#>  [25,]  280  150
#>  [26,]  282  156
#>  [27,]  282  162
#>  [28,]  275  162
#>  [29,]  269  161
#>  [30,]  262  159
#>  [31,]  259  161
#>  [32,]  258  168
#>  [33,]  252  171
#>  [34,]  246  168
#>  [35,]  239  163
#>  [36,]  233  158
#>  [37,]  226  151
#>  [38,]  226  153
#>  [39,]  229  160
#>  [40,]  231  167
#>  [41,]  232  173
#>  [42,]  233  180
#>  [43,]  233  186
#>  [44,]  230  192
#>  [45,]  223  191
#>  [46,]  217  189
#>  [47,]  211  188
#>  [48,]  210  194
#>  [49,]  210  201
#>  [50,]  208  208
#>  [51,]  206  214
#>  [52,]  202  221
#>  [53,]  196  226
#>  [54,]  190  220
#>  [55,]  186  213
#>  [56,]  184  207
#>  [57,]  182  200
#>  [58,]  181  194
#>  [59,]  181  187
#>  [60,]  174  189
#>  [61,]  168  191
#>  [62,]  161  192
#>  [63,]  159  186
#>  [64,]  159  179
#>  [65,]  160  173
#>  [66,]  162  166
#>  [67,]  164  159
#>  [68,]  166  153
#>  [69,]  165  152
#>  [70,]  159  158
#>  [71,]  152  163
#>  [72,]  145  168
#>  [73,]  139  171
#>  [74,]  134  167
#>  [75,]  133  161
#>  [76,]  129  160
#>  [77,]  122  161
#>  [78,]  116  162
#>  [79,]  109  162
#>  [80,]  109  156
#>  [81,]  113  149
#>  [82,]  117  142
#>  [83,]  122  136
#>  [84,]  127  129
#>  [85,]  123  127
#>  [86,]  116  124
#>  [87,]  115  119
#>  [88,]  122  112
#>  [89,]  129  107
#>  [90,]  135  103
#>  [91,]  142   99
#>  [92,]  140   94
#>  [93,]  133   87
#>  [94,]  127   80
#>  [95,]  126   75
#>  [96,]  133   72
#>  [97,]  139   70
#>  [98,]  146   68
#>  [99,]  153   66
#> [100,]  159   66
#> [101,]  166   65
#> [102,]  172   65
#> [103,]  179   66
#> [104,]  186   68
#> [105,]  192   70
#> [106,]  192   63
#> [107,]  191   57
#> [108,]  190   50
#> [109,]  189   44
#> [110,]  187   37
#> [111,]  183   30
#> [112,]  179   24
#> [113,]  184   20
#> [114,]  190   22
#> [115,]  194   29
#> [116,]  196   36
#> [117,]  197   42
#> [118,]  199   49
#> [119,]  200   55
#> [120,]  200   62
#> 
coo_reverse(bot)
#> # A tibble: 40 × 3
#>    coo             type   dummy
#>    <named list>    <fct>  <fct>
#>  1 <dbl [138 × 2]> whisky a    
#>  2 <dbl [168 × 2]> whisky a    
#>  3 <dbl [189 × 2]> whisky a    
#>  4 <dbl [129 × 2]> whisky a    
#>  5 <dbl [152 × 2]> whisky a    
#>  6 <dbl [161 × 2]> whisky a    
#>  7 <dbl [124 × 2]> whisky a    
#>  8 <dbl [126 × 2]> whisky a    
#>  9 <dbl [183 × 2]> whisky a    
#> 10 <dbl [193 × 2]> whisky a    
#> # ℹ 30 more rows