Create single shape made of coordinates

coo_single objects are tibble with exactly two columns named x and y. Anything that can be turned should work.

new_coo_single(x = tibble::tibble(x = double(), y = double()))

validate_coo_single(x)

coo_single(x)

is_coo_single(x)

is_coo_single1(x)

Arguments

x
anything that can be turned a tibble by tibble::tibble() should work

Value

a coo_single object

Details

You should use the helper coo_single, new_coo_single and validate_coo_single are for internal operations.

Functions

new_coo_single: Constructor

validate_coo_single: Validator

coo_single: Helper

is_coo_single: Class tester

is_coo_single1: Class1 tester

Examples

coo_single()

#> # A tibble: 0 x 2
#> # … with 2 variables: x <dbl>, y <dbl>
#> ❯coo_single with 0 coordinates

matrix(1:12, ncol=2) %>% coo_single()

#> # A tibble: 6 x 2
#>       x     y
#>   <int> <int>
#> 1     1     7
#> 2     2     8
#> 3     3     9
#> 4     4    10
#> 5     5    11
#> 6     6    12
#> ❯coo_single with 6 coordinates

list(x=1:2, y=3:4) %>% coo_single()

#> # A tibble: 2 x 2
#>       x     y
#>   <int> <int>
#> 1     1     3
#> 2     2     4
#> ❯coo_single with 2 coordinates