Untilt to axis

Rotate shape so the first point is exactly on an axis.

Usage

coo_untilt_x(x, ..., .cols = NULL, .ldk_col = NULL)

coo_untilt_y(x, ..., .cols = NULL, .ldk_col = NULL)

Arguments

x

A matrix (nx2), list of matrices, or tibble with coo columns.

...

Additional arguments (reserved for future use).

.cols

Column name(s) to process when x is a tibble. If NULL, automatically detects columns containing coo columns.

.ldk_col

Character. Name of landmark column. If NULL, uses colname_ldk.

Value

If x is a single matrix: returns the rotated matrix

If x is a list: returns a list of rotated matrices

If x is a tibble: returns the tibble with rotated coo column(s)

Details

coo_untilt_x(): Rotates around centroid so first point is on x-axis (y=0).

coo_untilt_y(): Rotates around centroid so first point is on y-axis (x=0), choosing the nearest y-axis (positive or negative).

Landmarks are synced.

Examples

coo_untilt_x(shapes$cat)
#>               [,1]     [,2]
#>   [1,]   11.031500 209.0988
#>   [2,]   15.276850 203.2698
#>   [3,]   16.660887 207.0227
#>   [4,]   20.613404 211.9601
#>   [5,]   27.280819 212.6987
#>   [6,]   30.587938 207.2158
#>   [7,]   28.857892 202.5246
#>   [8,]   27.720068 196.5492
#>   [9,]   25.644013 190.9198
#>  [10,]   23.567958 185.2905
#>  [11,]   20.553672 180.0071
#>  [12,]   18.477617 174.3777
#>  [13,]   16.401562 168.7483
#>  [14,]   14.325507 163.1189
#>  [15,]   12.249452 157.4895
#>  [16,]    9.235166 152.2061
#>  [17,]   11.011832 148.3534
#>  [18,]   15.356979 145.6852
#>  [19,]   22.024393 146.4238
#>  [20,]   28.445595 149.3849
#>  [21,]   32.398112 154.3223
#>  [22,]   36.350629 159.2596
#>  [23,]   39.364915 164.5430
#>  [24,]   42.379201 169.8264
#>  [25,]   44.455256 175.4558
#>  [26,]   45.593080 181.4312
#>  [27,]   46.730904 187.4066
#>  [28,]   47.868727 193.3820
#>  [29,]   49.598773 198.0731
#>  [30,]   54.143512 201.7263
#>  [31,]   55.820382 194.7128
#>  [32,]   55.620789 188.3914
#>  [33,]   55.421197 182.0700
#>  [34,]   54.283373 176.0946
#>  [35,]   53.145549 170.1192
#>  [36,]   51.069494 164.4898
#>  [37,]   48.993439 158.8605
#>  [38,]   45.979153 153.5771
#>  [39,]   42.964867 148.2937
#>  [40,]   39.358359 144.2946
#>  [41,]   34.467611 139.7032
#>  [42,]   27.700400 135.8039
#>  [43,]   20.686977 134.1270
#>  [44,]   14.019562 133.3883
#>  [45,]    8.390176 135.4644
#>  [46,]    2.760789 137.5405
#>  [47,]   -2.176580 141.4930
#>  [48,]   -6.767939 146.3837
#>  [49,]  -12.051317 149.3980
#>  [50,]  -17.680704 151.4741
#>  [51,]  -22.717868 152.2659
#>  [52,]  -28.693264 153.4037
#>  [53,]  -34.322651 155.4798
#>  [54,]  -40.298047 156.6176
#>  [55,]  -45.927434 158.6936
#>  [56,]  -51.556821 160.7697
#>  [57,]  -56.840198 163.7840
#>  [58,]  -62.469585 165.8600
#>  [59,]  -67.406954 169.8126
#>  [60,]  -71.998313 174.7033
#>  [61,]  -76.589672 179.5940
#>  [62,]  -80.588810 183.2006
#>  [63,]  -84.834160 189.0295
#>  [64,]  -88.733501 195.7967
#>  [65,]  -92.286833 203.5022
#>  [66,]  -93.025472 210.1696
#>  [67,]  -92.825879 216.4910
#>  [68,]  -93.564517 223.1584
#>  [69,]  -96.179618 230.5179
#>  [70,] -100.078959 237.2851
#>  [71,] -104.670319 242.1758
#>  [72,] -109.361474 243.9059
#>  [73,] -114.644852 246.9201
#>  [74,] -119.582220 250.8727
#>  [75,] -124.173580 255.7634
#>  [76,] -126.788680 263.1228
#>  [77,] -128.465550 270.1363
#>  [78,] -133.056909 275.0270
#>  [79,] -137.302260 280.8560
#>  [80,] -140.263370 287.2772
#>  [81,] -134.188177 289.3001
#>  [82,] -128.212781 288.1623
#>  [83,] -123.521626 286.4322
#>  [84,] -119.076683 286.9246
#>  [85,] -114.185934 291.5160
#>  [86,] -108.702964 294.8231
#>  [87,] -102.281763 297.7842
#>  [88,]  -96.798792 301.0913
#>  [89,]  -91.515415 298.0771
#>  [90,]  -87.270065 292.2481
#>  [91,]  -84.654964 284.8886
#>  [92,]  -81.001836 280.3439
#>  [93,]  -75.372449 278.2678
#>  [94,]  -70.335284 277.4760
#>  [95,]  -64.359888 276.3382
#>  [96,]  -58.730501 274.2622
#>  [97,]  -53.101114 272.1861
#>  [98,]  -47.471727 270.1100
#>  [99,]  -42.534359 266.1575
#> [100,]  -38.289009 260.3285
#> [101,]  -34.043659 254.4996
#> [102,]  -29.452299 249.6088
#> [103,]  -24.514931 245.6563
#> [104,]  -19.231553 242.6420
#> [105,]  -14.540397 240.9120
#> [106,]   -8.911010 238.8359
#> [107,]   -3.281624 236.7599
#> [108,]    2.347763 234.6838
#> [109,]    7.977150 232.6077
#> [110,]   13.952546 231.4699
#> [111,]   19.335720 231.6163
#> [112,]   23.288237 236.5537
#> [113,]   28.671411 236.7001
#> [114,]   33.608780 232.7476
#> [115,]   35.285649 225.7342
#> [116,]   33.555603 221.0430
#> [117,]   29.603086 216.1057
#> [118,]   22.243654 213.4906
#> [119,]   15.230230 211.8137
#> [120,]    8.908825 212.0133
#> attr(,"class")
#> [1] "coo"    "matrix" "array" 
coo_untilt_y(shapes$cat)
#>            [,1]       [,2]
#>   [1,] 209.0988 -11.031500
#>   [2,] 203.2698 -15.276850
#>   [3,] 207.0227 -16.660887
#>   [4,] 211.9601 -20.613404
#>   [5,] 212.6987 -27.280819
#>   [6,] 207.2158 -30.587938
#>   [7,] 202.5246 -28.857892
#>   [8,] 196.5492 -27.720068
#>   [9,] 190.9198 -25.644013
#>  [10,] 185.2905 -23.567958
#>  [11,] 180.0071 -20.553672
#>  [12,] 174.3777 -18.477617
#>  [13,] 168.7483 -16.401562
#>  [14,] 163.1189 -14.325507
#>  [15,] 157.4895 -12.249452
#>  [16,] 152.2061  -9.235166
#>  [17,] 148.3534 -11.011832
#>  [18,] 145.6852 -15.356979
#>  [19,] 146.4238 -22.024393
#>  [20,] 149.3849 -28.445595
#>  [21,] 154.3223 -32.398112
#>  [22,] 159.2596 -36.350629
#>  [23,] 164.5430 -39.364915
#>  [24,] 169.8264 -42.379201
#>  [25,] 175.4558 -44.455256
#>  [26,] 181.4312 -45.593080
#>  [27,] 187.4066 -46.730904
#>  [28,] 193.3820 -47.868727
#>  [29,] 198.0731 -49.598773
#>  [30,] 201.7263 -54.143512
#>  [31,] 194.7128 -55.820382
#>  [32,] 188.3914 -55.620789
#>  [33,] 182.0700 -55.421197
#>  [34,] 176.0946 -54.283373
#>  [35,] 170.1192 -53.145549
#>  [36,] 164.4898 -51.069494
#>  [37,] 158.8605 -48.993439
#>  [38,] 153.5771 -45.979153
#>  [39,] 148.2937 -42.964867
#>  [40,] 144.2946 -39.358359
#>  [41,] 139.7032 -34.467611
#>  [42,] 135.8039 -27.700400
#>  [43,] 134.1270 -20.686977
#>  [44,] 133.3883 -14.019562
#>  [45,] 135.4644  -8.390176
#>  [46,] 137.5405  -2.760789
#>  [47,] 141.4930   2.176580
#>  [48,] 146.3837   6.767939
#>  [49,] 149.3980  12.051317
#>  [50,] 151.4741  17.680704
#>  [51,] 152.2659  22.717868
#>  [52,] 153.4037  28.693264
#>  [53,] 155.4798  34.322651
#>  [54,] 156.6176  40.298047
#>  [55,] 158.6936  45.927434
#>  [56,] 160.7697  51.556821
#>  [57,] 163.7840  56.840198
#>  [58,] 165.8600  62.469585
#>  [59,] 169.8126  67.406954
#>  [60,] 174.7033  71.998313
#>  [61,] 179.5940  76.589672
#>  [62,] 183.2006  80.588810
#>  [63,] 189.0295  84.834160
#>  [64,] 195.7967  88.733501
#>  [65,] 203.5022  92.286833
#>  [66,] 210.1696  93.025472
#>  [67,] 216.4910  92.825879
#>  [68,] 223.1584  93.564517
#>  [69,] 230.5179  96.179618
#>  [70,] 237.2851 100.078959
#>  [71,] 242.1758 104.670319
#>  [72,] 243.9059 109.361474
#>  [73,] 246.9201 114.644852
#>  [74,] 250.8727 119.582220
#>  [75,] 255.7634 124.173580
#>  [76,] 263.1228 126.788680
#>  [77,] 270.1363 128.465550
#>  [78,] 275.0270 133.056909
#>  [79,] 280.8560 137.302260
#>  [80,] 287.2772 140.263370
#>  [81,] 289.3001 134.188177
#>  [82,] 288.1623 128.212781
#>  [83,] 286.4322 123.521626
#>  [84,] 286.9246 119.076683
#>  [85,] 291.5160 114.185934
#>  [86,] 294.8231 108.702964
#>  [87,] 297.7842 102.281763
#>  [88,] 301.0913  96.798792
#>  [89,] 298.0771  91.515415
#>  [90,] 292.2481  87.270065
#>  [91,] 284.8886  84.654964
#>  [92,] 280.3439  81.001836
#>  [93,] 278.2678  75.372449
#>  [94,] 277.4760  70.335284
#>  [95,] 276.3382  64.359888
#>  [96,] 274.2622  58.730501
#>  [97,] 272.1861  53.101114
#>  [98,] 270.1100  47.471727
#>  [99,] 266.1575  42.534359
#> [100,] 260.3285  38.289009
#> [101,] 254.4996  34.043659
#> [102,] 249.6088  29.452299
#> [103,] 245.6563  24.514931
#> [104,] 242.6420  19.231553
#> [105,] 240.9120  14.540397
#> [106,] 238.8359   8.911010
#> [107,] 236.7599   3.281624
#> [108,] 234.6838  -2.347763
#> [109,] 232.6077  -7.977150
#> [110,] 231.4699 -13.952546
#> [111,] 231.6163 -19.335720
#> [112,] 236.5537 -23.288237
#> [113,] 236.7001 -28.671411
#> [114,] 232.7476 -33.608780
#> [115,] 225.7342 -35.285649
#> [116,] 221.0430 -33.555603
#> [117,] 216.1057 -29.603086
#> [118,] 213.4906 -22.243654
#> [119,] 211.8137 -15.230230
#> [120,] 212.0133  -8.908825
#> attr(,"class")
#> [1] "coo"    "matrix" "array"