Principal component analysis

Performs a principal components analysis (delegated to stats::prcomp).

stat_pca(x, center = TRUE, scale = TRUE)

Arguments

x
coe_tbl

center
logical whether the variables should be shifted to be zero centered (center in stats::prcomp)

scale
logical whether the variables should be scaled to have unit variance before the analysis takes place (scale. in stats::prcomp)

Value

object of class pca, a tibble::tibble with attributes

References

stats::prcomp

Examples

(x <-  bot2 %>% efourier(4) %>% stat_pca())

#> ! describe_pca: 'center' was missing, set to TRUE

#> ! describe_pca: 'scale' was missing, set to TRUE

#> # A tibble: 40 x 19
#>    type  fake  coe      PC1     PC2    PC3    PC4      PC5     PC6     PC7
#>    <fct> <fct> <nam>  <dbl>   <dbl>  <dbl>  <dbl>    <dbl>   <dbl>   <dbl>
#>  1 whis… a     <coe… -1.29  -4.45    1.83   1.85   0.228    1.45   -0.811 
#>  2 whis… a     <coe… -1.47  -2.15   -0.887  1.17   0.276   -0.531   0.126 
#>  3 whis… a     <coe…  0.757  1.60   -0.438 -0.637  0.744    1.06   -0.374 
#>  4 whis… a     <coe…  1.12  -0.516  -0.902  1.83  -0.00625 -0.220  -0.176 
#>  5 whis… a     <coe… -2.50   0.0999 -3.53  -0.265  0.603    0.236  -0.334 
#>  6 whis… a     <coe… -1.46   2.15    1.34   0.622  1.21    -0.0776  0.309 
#>  7 whis… a     <coe…  0.356 -1.41    1.08   1.56  -0.263   -0.0261  0.268 
#>  8 whis… a     <coe…  0.961  0.128  -1.50  -0.787  1.03    -0.155  -0.0694
#>  9 whis… a     <coe…  1.79   0.376  -0.673 -1.19   0.378    0.341   0.201 
#> 10 whis… a     <coe… -0.976 -2.57   -1.58   0.455  0.391    0.575   0.0777
#> # … with 30 more rows, and 9 more variables: PC8 <dbl>, PC9 <dbl>, PC10 <dbl>,
#> #   PC11 <dbl>, PC12 <dbl>, PC13 <dbl>, PC14 <dbl>, PC15 <dbl>, PC16 <dbl>
#> ❯ a stat_pca with 19 shapes

attributes(x)

#> $class
#> [1] "pca"        "coe_tbl"    "tbl_df"     "tbl"        "data.frame"
#> 
#> $row.names
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40
#> 
#> $names
#>  [1] "type" "fake" "coe"  "PC1"  "PC2"  "PC3"  "PC4"  "PC5"  "PC6"  "PC7" 
#> [11] "PC8"  "PC9"  "PC10" "PC11" "PC12" "PC13" "PC14" "PC15" "PC16"
#> 
#> $method
#> [1] "prcomp"
#> 
#> $coe_class
#>        coe 
#> "efourier" 
#> 
#> $center
#>                PC1          PC2           PC3          PC4         PC5
#> coe_a1  0.26106944  0.262685299 -1.267944e-01  0.312081484 -0.32486653
#> coe_a2 -0.33612538  0.228661341 -5.396331e-03  0.035778566 -0.14646125
#> coe_a3 -0.10435611 -0.407975401 -2.083467e-01  0.105644800  0.46356291
#> coe_a4 -0.04568896  0.173582117 -5.677093e-01 -0.003103513 -0.04698667
#> coe_b1  0.30838363 -0.048734436 -2.283138e-01 -0.042576582 -0.25686069
#> coe_b2  0.30641842  0.019834948  2.858074e-01  0.307897944  0.07031098
#> coe_b3 -0.29658304 -0.138880585  9.893450e-02  0.382630669  0.29068565
#> coe_b4  0.14280844 -0.246709709  4.623740e-01 -0.169679201 -0.10157647
#> coe_c1 -0.31645209 -0.095511415  2.441745e-01  0.275113224 -0.19496129
#> coe_c2  0.30401660 -0.163200727  2.267035e-01 -0.035727588 -0.19529198
#> coe_c3  0.25803081  0.215088703  6.372147e-05 -0.349037835  0.52970642
#> coe_c4 -0.24424177 -0.003302629  6.382634e-02 -0.612913550 -0.14619649
#> coe_d1  0.12919889  0.465335019  1.039691e-01  0.091123467  0.11533742
#> coe_d2  0.28476080 -0.255450351  5.130087e-03 -0.089335803  0.02478813
#> coe_d3 -0.04105124 -0.433499265 -2.766249e-01 -0.045375296 -0.29533597
#> coe_d4 -0.30044651  0.217875280  2.415665e-01 -0.159831007 -0.10451732
#>                PC6           PC7         PC8         PC9         PC10
#> coe_a1  0.07920565 -0.0072735930 -0.02106663  0.58452089 -0.482908606
#> coe_a2 -0.12368415  0.0617605047 -0.34387565 -0.20425826 -0.307518801
#> coe_a3  0.09698571 -0.1577397729 -0.27248569  0.18570781 -0.172991119
#> coe_a4 -0.26998335  0.0102511899 -0.18210801  0.04057531  0.333018535
#> coe_b1  0.49469919 -0.5256878567 -0.17541572 -0.32782051  0.112979734
#> coe_b2  0.04182700  0.0556869818  0.09047772  0.28015208  0.582383939
#> coe_b3  0.00189172 -0.4534068852 -0.09467136  0.20780565 -0.006156167
#> coe_b4  0.25613038  0.0979615622 -0.10738412  0.01290381 -0.224782051
#> coe_c1 -0.07809167  0.0158564180  0.10177987 -0.21957233  0.028793005
#> coe_c2 -0.40735565 -0.1347027406 -0.63335888 -0.02321990  0.065993053
#> coe_c3  0.08267212  0.2284127679 -0.20653266  0.06758065 -0.086093248
#> coe_c4 -0.01693182 -0.4257622070  0.21949756  0.37829634  0.050295546
#> coe_d1 -0.10467302 -0.3241538512  0.08804721 -0.04289224 -0.040477764
#> coe_d2 -0.62571812 -0.2122154673  0.27688677 -0.04165935 -0.181023068
#> coe_d3  0.04883616  0.2730735801 -0.03102156  0.27038828  0.105407192
#> coe_d4 -0.00504455  0.0006087129 -0.35598397  0.27940555  0.261265316
#>                PC11         PC12         PC13        PC14        PC15
#> coe_a1  0.127007744 -0.014410319  0.109997047 -0.01863276 -0.17326370
#> coe_a2 -0.229472605  0.174031538 -0.367948144  0.54935678 -0.13181600
#> coe_a3  0.530569073  0.149769201 -0.184649405  0.14140732  0.09636102
#> coe_a4  0.115554087 -0.528047960 -0.248213227 -0.09756235 -0.23976821
#> coe_b1 -0.039790608  0.197668733 -0.026032072 -0.01435578 -0.26046811
#> coe_b2  0.018928753  0.093939727 -0.237523518  0.45265440 -0.14407570
#> coe_b3 -0.550622652 -0.209303488  0.097122505 -0.13622466 -0.12708880
#> coe_b4 -0.004912152 -0.493096470 -0.480029895 -0.15049884 -0.11400469
#> coe_c1  0.352220402 -0.073841781  0.339991722  0.06327524 -0.47638375
#> coe_c2  0.031560068 -0.123937183  0.346150487  0.10019676  0.23081374
#> coe_c3 -0.186881880  0.027958374  0.278858056  0.08026068 -0.43832749
#> coe_c4  0.085303488 -0.136192526  0.113333634  0.35310175 -0.02229967
#> coe_d1  0.016043692  0.001718047 -0.171508891 -0.04219763  0.38222063
#> coe_d2 -0.065037188  0.282998316 -0.271351442 -0.14841535 -0.34251472
#> coe_d3 -0.382834254  0.185777911  0.000976822  0.03330662  0.10894350
#> coe_d4  0.113569234  0.422661432 -0.175454596 -0.50194969 -0.14127123
#>                 PC16
#> coe_a1 -5.820779e-02
#> coe_a2 -2.868687e-02
#> coe_a3  1.137763e-01
#> coe_a4  6.131066e-02
#> coe_b1  3.386887e-02
#> coe_b2 -4.833727e-02
#> coe_b3 -5.619943e-02
#> coe_b4  1.384686e-01
#> coe_c1  4.190655e-01
#> coe_c2 -2.274605e-02
#> coe_c3  2.430176e-01
#> coe_c4  7.149859e-05
#> coe_d1  6.553958e-01
#> coe_d2 -3.479953e-03
#> coe_d3  5.370272e-01
#> coe_d4 -1.116285e-02
#> 
#> $scale
#>                PC1          PC2           PC3          PC4         PC5
#> coe_a1  0.26106944  0.262685299 -1.267944e-01  0.312081484 -0.32486653
#> coe_a2 -0.33612538  0.228661341 -5.396331e-03  0.035778566 -0.14646125
#> coe_a3 -0.10435611 -0.407975401 -2.083467e-01  0.105644800  0.46356291
#> coe_a4 -0.04568896  0.173582117 -5.677093e-01 -0.003103513 -0.04698667
#> coe_b1  0.30838363 -0.048734436 -2.283138e-01 -0.042576582 -0.25686069
#> coe_b2  0.30641842  0.019834948  2.858074e-01  0.307897944  0.07031098
#> coe_b3 -0.29658304 -0.138880585  9.893450e-02  0.382630669  0.29068565
#> coe_b4  0.14280844 -0.246709709  4.623740e-01 -0.169679201 -0.10157647
#> coe_c1 -0.31645209 -0.095511415  2.441745e-01  0.275113224 -0.19496129
#> coe_c2  0.30401660 -0.163200727  2.267035e-01 -0.035727588 -0.19529198
#> coe_c3  0.25803081  0.215088703  6.372147e-05 -0.349037835  0.52970642
#> coe_c4 -0.24424177 -0.003302629  6.382634e-02 -0.612913550 -0.14619649
#> coe_d1  0.12919889  0.465335019  1.039691e-01  0.091123467  0.11533742
#> coe_d2  0.28476080 -0.255450351  5.130087e-03 -0.089335803  0.02478813
#> coe_d3 -0.04105124 -0.433499265 -2.766249e-01 -0.045375296 -0.29533597
#> coe_d4 -0.30044651  0.217875280  2.415665e-01 -0.159831007 -0.10451732
#>                PC6           PC7         PC8         PC9         PC10
#> coe_a1  0.07920565 -0.0072735930 -0.02106663  0.58452089 -0.482908606
#> coe_a2 -0.12368415  0.0617605047 -0.34387565 -0.20425826 -0.307518801
#> coe_a3  0.09698571 -0.1577397729 -0.27248569  0.18570781 -0.172991119
#> coe_a4 -0.26998335  0.0102511899 -0.18210801  0.04057531  0.333018535
#> coe_b1  0.49469919 -0.5256878567 -0.17541572 -0.32782051  0.112979734
#> coe_b2  0.04182700  0.0556869818  0.09047772  0.28015208  0.582383939
#> coe_b3  0.00189172 -0.4534068852 -0.09467136  0.20780565 -0.006156167
#> coe_b4  0.25613038  0.0979615622 -0.10738412  0.01290381 -0.224782051
#> coe_c1 -0.07809167  0.0158564180  0.10177987 -0.21957233  0.028793005
#> coe_c2 -0.40735565 -0.1347027406 -0.63335888 -0.02321990  0.065993053
#> coe_c3  0.08267212  0.2284127679 -0.20653266  0.06758065 -0.086093248
#> coe_c4 -0.01693182 -0.4257622070  0.21949756  0.37829634  0.050295546
#> coe_d1 -0.10467302 -0.3241538512  0.08804721 -0.04289224 -0.040477764
#> coe_d2 -0.62571812 -0.2122154673  0.27688677 -0.04165935 -0.181023068
#> coe_d3  0.04883616  0.2730735801 -0.03102156  0.27038828  0.105407192
#> coe_d4 -0.00504455  0.0006087129 -0.35598397  0.27940555  0.261265316
#>                PC11         PC12         PC13        PC14        PC15
#> coe_a1  0.127007744 -0.014410319  0.109997047 -0.01863276 -0.17326370
#> coe_a2 -0.229472605  0.174031538 -0.367948144  0.54935678 -0.13181600
#> coe_a3  0.530569073  0.149769201 -0.184649405  0.14140732  0.09636102
#> coe_a4  0.115554087 -0.528047960 -0.248213227 -0.09756235 -0.23976821
#> coe_b1 -0.039790608  0.197668733 -0.026032072 -0.01435578 -0.26046811
#> coe_b2  0.018928753  0.093939727 -0.237523518  0.45265440 -0.14407570
#> coe_b3 -0.550622652 -0.209303488  0.097122505 -0.13622466 -0.12708880
#> coe_b4 -0.004912152 -0.493096470 -0.480029895 -0.15049884 -0.11400469
#> coe_c1  0.352220402 -0.073841781  0.339991722  0.06327524 -0.47638375
#> coe_c2  0.031560068 -0.123937183  0.346150487  0.10019676  0.23081374
#> coe_c3 -0.186881880  0.027958374  0.278858056  0.08026068 -0.43832749
#> coe_c4  0.085303488 -0.136192526  0.113333634  0.35310175 -0.02229967
#> coe_d1  0.016043692  0.001718047 -0.171508891 -0.04219763  0.38222063
#> coe_d2 -0.065037188  0.282998316 -0.271351442 -0.14841535 -0.34251472
#> coe_d3 -0.382834254  0.185777911  0.000976822  0.03330662  0.10894350
#> coe_d4  0.113569234  0.422661432 -0.175454596 -0.50194969 -0.14127123
#>                 PC16
#> coe_a1 -5.820779e-02
#> coe_a2 -2.868687e-02
#> coe_a3  1.137763e-01
#> coe_a4  6.131066e-02
#> coe_b1  3.386887e-02
#> coe_b2 -4.833727e-02
#> coe_b3 -5.619943e-02
#> coe_b4  1.384686e-01
#> coe_c1  4.190655e-01
#> coe_c2 -2.274605e-02
#> coe_c3  2.430176e-01
#> coe_c4  7.149859e-05
#> coe_d1  6.553958e-01
#> coe_d2 -3.479953e-03
#> coe_d3  5.370272e-01
#> coe_d4 -1.116285e-02
#> 
#> $rotation
#>                PC1          PC2           PC3          PC4         PC5
#> coe_a1  0.26106944  0.262685299 -1.267944e-01  0.312081484 -0.32486653
#> coe_a2 -0.33612538  0.228661341 -5.396331e-03  0.035778566 -0.14646125
#> coe_a3 -0.10435611 -0.407975401 -2.083467e-01  0.105644800  0.46356291
#> coe_a4 -0.04568896  0.173582117 -5.677093e-01 -0.003103513 -0.04698667
#> coe_b1  0.30838363 -0.048734436 -2.283138e-01 -0.042576582 -0.25686069
#> coe_b2  0.30641842  0.019834948  2.858074e-01  0.307897944  0.07031098
#> coe_b3 -0.29658304 -0.138880585  9.893450e-02  0.382630669  0.29068565
#> coe_b4  0.14280844 -0.246709709  4.623740e-01 -0.169679201 -0.10157647
#> coe_c1 -0.31645209 -0.095511415  2.441745e-01  0.275113224 -0.19496129
#> coe_c2  0.30401660 -0.163200727  2.267035e-01 -0.035727588 -0.19529198
#> coe_c3  0.25803081  0.215088703  6.372147e-05 -0.349037835  0.52970642
#> coe_c4 -0.24424177 -0.003302629  6.382634e-02 -0.612913550 -0.14619649
#> coe_d1  0.12919889  0.465335019  1.039691e-01  0.091123467  0.11533742
#> coe_d2  0.28476080 -0.255450351  5.130087e-03 -0.089335803  0.02478813
#> coe_d3 -0.04105124 -0.433499265 -2.766249e-01 -0.045375296 -0.29533597
#> coe_d4 -0.30044651  0.217875280  2.415665e-01 -0.159831007 -0.10451732
#>                PC6           PC7         PC8         PC9         PC10
#> coe_a1  0.07920565 -0.0072735930 -0.02106663  0.58452089 -0.482908606
#> coe_a2 -0.12368415  0.0617605047 -0.34387565 -0.20425826 -0.307518801
#> coe_a3  0.09698571 -0.1577397729 -0.27248569  0.18570781 -0.172991119
#> coe_a4 -0.26998335  0.0102511899 -0.18210801  0.04057531  0.333018535
#> coe_b1  0.49469919 -0.5256878567 -0.17541572 -0.32782051  0.112979734
#> coe_b2  0.04182700  0.0556869818  0.09047772  0.28015208  0.582383939
#> coe_b3  0.00189172 -0.4534068852 -0.09467136  0.20780565 -0.006156167
#> coe_b4  0.25613038  0.0979615622 -0.10738412  0.01290381 -0.224782051
#> coe_c1 -0.07809167  0.0158564180  0.10177987 -0.21957233  0.028793005
#> coe_c2 -0.40735565 -0.1347027406 -0.63335888 -0.02321990  0.065993053
#> coe_c3  0.08267212  0.2284127679 -0.20653266  0.06758065 -0.086093248
#> coe_c4 -0.01693182 -0.4257622070  0.21949756  0.37829634  0.050295546
#> coe_d1 -0.10467302 -0.3241538512  0.08804721 -0.04289224 -0.040477764
#> coe_d2 -0.62571812 -0.2122154673  0.27688677 -0.04165935 -0.181023068
#> coe_d3  0.04883616  0.2730735801 -0.03102156  0.27038828  0.105407192
#> coe_d4 -0.00504455  0.0006087129 -0.35598397  0.27940555  0.261265316
#>                PC11         PC12         PC13        PC14        PC15
#> coe_a1  0.127007744 -0.014410319  0.109997047 -0.01863276 -0.17326370
#> coe_a2 -0.229472605  0.174031538 -0.367948144  0.54935678 -0.13181600
#> coe_a3  0.530569073  0.149769201 -0.184649405  0.14140732  0.09636102
#> coe_a4  0.115554087 -0.528047960 -0.248213227 -0.09756235 -0.23976821
#> coe_b1 -0.039790608  0.197668733 -0.026032072 -0.01435578 -0.26046811
#> coe_b2  0.018928753  0.093939727 -0.237523518  0.45265440 -0.14407570
#> coe_b3 -0.550622652 -0.209303488  0.097122505 -0.13622466 -0.12708880
#> coe_b4 -0.004912152 -0.493096470 -0.480029895 -0.15049884 -0.11400469
#> coe_c1  0.352220402 -0.073841781  0.339991722  0.06327524 -0.47638375
#> coe_c2  0.031560068 -0.123937183  0.346150487  0.10019676  0.23081374
#> coe_c3 -0.186881880  0.027958374  0.278858056  0.08026068 -0.43832749
#> coe_c4  0.085303488 -0.136192526  0.113333634  0.35310175 -0.02229967
#> coe_d1  0.016043692  0.001718047 -0.171508891 -0.04219763  0.38222063
#> coe_d2 -0.065037188  0.282998316 -0.271351442 -0.14841535 -0.34251472
#> coe_d3 -0.382834254  0.185777911  0.000976822  0.03330662  0.10894350
#> coe_d4  0.113569234  0.422661432 -0.175454596 -0.50194969 -0.14127123
#>                 PC16
#> coe_a1 -5.820779e-02
#> coe_a2 -2.868687e-02
#> coe_a3  1.137763e-01
#> coe_a4  6.131066e-02
#> coe_b1  3.386887e-02
#> coe_b2 -4.833727e-02
#> coe_b3 -5.619943e-02
#> coe_b4  1.384686e-01
#> coe_c1  4.190655e-01
#> coe_c2 -2.274605e-02
#> coe_c3  2.430176e-01
#> coe_c4  7.149859e-05
#> coe_d1  6.553958e-01
#> coe_d2 -3.479953e-03
#> coe_d3  5.370272e-01
#> coe_d4 -1.116285e-02
#> 
#> $sdev
#>  [1] 2.58434299 1.93835626 1.59761175 1.18524202 0.82273809 0.67950868
#>  [7] 0.43145796 0.32387439 0.24455355 0.22818777 0.17663316 0.12754992
#> [13] 0.09181020 0.07402419 0.05538343 0.02798951
#> 
#> $mshape
#>      coe_a1      coe_a2      coe_a3      coe_a4      coe_b1      coe_b2 
#> -118.869287    7.907727   14.850701   -3.817288  -16.003513  -27.424758 
#>      coe_b3      coe_b4      coe_c1      coe_c2      coe_c3      coe_c4 
#>    6.417212    6.346127   48.474250   -3.163270  -12.364630    6.875723 
#>      coe_d1      coe_d2      coe_d3      coe_d4 
#> -357.327164   -1.096749   29.257028    4.548746 
#>